# CUsersSHAYANAppDataLocalTempopenapi.json
{"openapi":"3.1.0","info":{"title":"RAG Chatbot Backend","description":"A FastAPI backend for a RAG chatbot using LangChain, Qdrant, and Neon.","version":"0.1.0"},"paths":{"/ingest":{"post":{"summary":"Ingest Endpoint","description":"Endpoint to ingest a text file. It reads the file, splits it into chunks,\nand stores the embeddings in the specified Qdrant collection.","operationId":"ingest_endpoint_ingest_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_ingest_endpoint_ingest_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"summary":"Chat Endpoint","description":"Handles chat requests by first consulting the book (RAG) and then falling back to a general LLM.","operationId":"chat_endpoint_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat-selection":{"post":{"summary":"Chat Selection Endpoint","description":"Endpoint for asking questions based on user-selected text.\nThis bypasses the retriever and uses the provided text as context.","operationId":"chat_selection_endpoint_chat_selection_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatSelectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat-general":{"post":{"summary":"Chat General Endpoint","description":"General chat endpoint without RAG - for greetings and general questions","operationId":"chat_general_endpoint_chat_general_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_ingest_endpoint_ingest_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"collection_name":{"type":"string","title":"Collection Name"}},"type":"object","required":["file","collection_name"],"title":"Body_ingest_endpoint_ingest_post"},"ChatRequest":{"properties":{"collection_name":{"type":"string","title":"Collection Name"},"question":{"type":"string","title":"Question"}},"type":"object","required":["collection_name","question"],"title":"ChatRequest"},"ChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"}},"type":"object","required":["answer"],"title":"ChatResponse"},"ChatSelectionRequest":{"properties":{"question":{"type":"string","title":"Question"},"selected_text":{"type":"string","title":"Selected Text"}},"type":"object","required":["question","selected_text"],"title":"ChatSelectionRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}